Reads the table of pies mayores

Reads the table of pies mayores

Reads the table of pies menores

readPiesMayoresIFN2(
  prov,
  DBFdir = "DBF",
  plotIDs = NULL,
  rem.nodensity = FALSE,
  rem.nospecies = FALSE,
  rem.noheight = FALSE,
  height.cm = FALSE,
  subsetVars = TRUE
)

readPiesMayoresIFN3(
  accessFiles,
  plotTypeIFN3 = c("A1", "NN"),
  plotIDs = NULL,
  include.IFN2 = FALSE,
  rem.nospecies = FALSE,
  rem.noheight = FALSE,
  rem.nodensity = FALSE,
  rem.codes32 = character(0),
  rem.codes33 = character(0),
  height.cm = FALSE,
  subsetVars = TRUE
)

readPiesMayoresIFN4(
  accessFiles,
  plotTypeIFN4 = c("A1", "NN"),
  plotIDs = NULL,
  include.IFN3 = FALSE,
  rem.nospecies = FALSE,
  rem.noheight = FALSE,
  rem.nodensity = FALSE,
  rem.codes43 = character(0),
  rem.codes44 = character(0),
  height.cm = FALSE,
  subsetVars = TRUE,
  vars.IFN3 = c("OrdenIf3", "Especie", "Dn1", "Dn2", "Ht", "Forma")
)

readPiesMenoresIFN2(prov, DBFdir = "DBF", subsetVars = TRUE)

Arguments

prov

Character vector with codes of provinces (e.g. c("01","03")) or "all" for all provinces.

DBFdir

String with the path where provinces are stored

plotIDs

String vector with a selection of plots

rem.nodensity

Flag to remove records without density

rem.nospecies

Flag to remove records without species identity

rem.noheight

Flag to remove records without tree height

height.cm

Flag to return height in 'cm' instead of 'meters'

subsetVars

Flag to filter data columns

accessFiles

A character vector of access files to be read.

plotTypeIFN3

Subset of plot types to include:

N - Se levantan por primera vez en el IFN-3 y son utilizadas para el calculo de existencias actuales.

A - Parcelas levantadas en el IFN-2 que volveran a ser apeadas en el IFN-3. Existen cinco casos:

A1 . Parcelas en las que se localice el rejon y esta bien implantado. Se utilizan para la comparacion de inventarios y para el calculo de existencias actuales.

A3 . Parcela apeada en IFN2, pinchazo IFN2 y coordenada UTM situadas en diferentes teselas. Se cumplimentan dos estadillos:

A3C . Parcela para la Comparacion de Inventarios. Se levantan en el punto donde se encuentra el rejon del IFN-2.

A3E. Parcelas para el calculo de Existencias Actuales. Se levantan en el emplazamiento del IFN-3 ( Coordenada UTM)

A4 . Parcelas en las que no se localizo el rejon. Se utilizan para el calculo de existencias actuales.

A4C . Parcelas en las que no se localice el rej?n por cambio totales en la cubierta forestal de la tesela que contiene la parcela (p?rdida total de la masa muestreada en el IFN2 o incorporaci?n de una nueva masa forestal inexistente en el IFN2). Se utilizar?n para la comparaci?n de inventarios y para el c?lculo de existencias actuales.

A6C . Parcelas en las que no se localice el rej?n por cambios totales en la cubierta forestal de la tesela que contiene la parcela (p?rdida total de la masa muestreada en el IFN2 e incorporaci?n de una nueva masa forestal). Se utilizar?n para la comparaci?n de inventarios y para el c?lculo de existencias actuales.

include.IFN2

Flag to include diameter and height measured in IFN2

rem.codes32

set of tree codes in IFN2 corresponding to records to be removed:

'0' or '000' - Tree not tallied in IFN2 because of its small size.

'999' - Tree that should have been tallied in IFN2 but was not.

rem.codes33

set of tree codes in IFN3 corresponding to records to be removed

'000' - Tree tallied in IFN2 but cut between the two inventories.

'444' - Tree tallied in IFN2 and the stump is not found or it is in advanced decomposition state.

'777' - Cork oak tree tallied in IFN2 and scorched between the two inventories.

'888' - Tree tallied in IFN2 and dead between the two inventories.

'999' - Tree tallied in IFN2 but should have not been included.

plotTypeIFN4

Subset of plot types to include:

include.IFN3

Flag to include diameter and height measured in IFN3

rem.codes43

set of tree codes in IFN3 corresponding to records to be removed:

'0' or '000' - Tree not tallied in IFN3 because of its small size.

'999' - Tree that should have been tallied in IFN3 but was not for some reason.

rem.codes44

set of tree codes in IFN4 corresponding to records to be removed

'000' - Tree tallied in IFN3 but cut between the two inventories (stump is found).

'444' - Tree tallied in IFN3 and the stump is not found or it is in advanced decomposition state.

'777' - Cork oak tree tallied in IFN3 and scorched between the two inventories.

'888' - Tree tallied in IFN3 and dead between the two inventories.

'999' - Tree tallied in IFN3 but should have not been included.

vars.IFN3

String vector with column names of variables in the IFN3 table

Value

a data frame

a data frame

Examples

if (FALSE) {
accessfilesIFN3 = list.files("//SERVERPROCESS/Miquel/Datasets/IFN/IFN3/MSAccess/",
                             pattern = "*[.]accdb", full.names = T)

treeData = readPiesMayoresIFN3(accessfilesIFN3[prov])

}